refactor(input): rename with Wrapper
