+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/StandardField/StandardFieldValue/DSDatePickerWrapper.tsx
@@ -1,6 +1,3 @@
-// Copyright LEAV Solutions 2017
-// This file is released under LGPL V3
-// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
 import {KitDatePicker} from 'aristid-ds';
 import {FunctionComponent} from 'react';
 import {
@@ -11,14 +8,14 @@ import {Form} from 'antd';
 import dayjs from 'dayjs';
 import styled from 'styled-components';
 import {IProvidedByAntFormItem, StandardValueTypes} from '../../../_types';
-import {RangePickerProps} from 'antd/lib/date-picker';
+import {DatePickerProps} from 'antd/lib/date-picker';
 import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import {RecordFormAttributeFragment} from '_ui/_gqlTypes';
 import {useValueDetailsButton} from '_ui/components/RecordEdition/EditRecordContent/shared/ValueDetailsBtn/useValueDetailsButton';
 import {useLang} from '_ui/hooks';
 import {localizedTranslation} from '@leav/utils';
 
-interface IDSRangePickerProps extends IProvidedByAntFormItem<RangePickerProps> {
+interface IDSDatePickerWrapperProps extends IProvidedByAntFormItem<DatePickerProps> {
     state: IStandardFieldReducerState;
     attribute: RecordFormAttributeFragment;
     fieldValue: IStandardFieldValue;
@@ -26,11 +23,11 @@ interface IDSRangePickerProps extends IProvidedByAntFormItem<RangePickerProps> {
     shouldShowValueDetailsButton?: boolean;
 }
 
-const KitDatePickerRangePickerStyled = styled(KitDatePicker.RangePicker)<{$shouldHighlightColor: boolean}>`
+const KitDatePickerStyled = styled(KitDatePicker)<{$shouldHighlightColor: boolean}>`
     color: ${({$shouldHighlightColor}) => ($shouldHighlightColor ? 'var(--general-colors-primary-400)' : 'initial')};
 `;
 
-export const DSRangePicker: FunctionComponent<IDSRangePickerProps> = ({
+export const DSDatePickerWrapper: FunctionComponent<IDSDatePickerWrapperProps> = ({
     value,
     onChange,
     state,
@@ -47,43 +44,35 @@ export const DSRangePicker: FunctionComponent<IDSRangePickerProps> = ({
         attribute
     });
 
-    const _handleDateChange: (
-        rangePickerDates: [from: dayjs.Dayjs, to: dayjs.Dayjs] | null,
-        antOnChangeParams: [from: string, to: string] | null
-    ) => void = (rangePickerDates, ...antOnChangeParams) => {
-        if (state.isInheritedValue && rangePickerDates === null) {
-            onChange(
-                [
-                    dayjs.unix(Number(state.inheritedValue.raw_value.from)),
-                    dayjs.unix(Number(state.inheritedValue.raw_value.to))
-                ],
-                state.inheritedValue.raw_value
-            );
+    const _handleDateChange: (datePickerDate: dayjs.Dayjs | null, antOnChangeParams: string | string[]) => void = (
+        datePickerDate,
+        ...antOnChangeParams
+    ) => {
+        if (state.isInheritedValue && datePickerDate === null) {
+            onChange(dayjs.unix(Number(state.inheritedValue.raw_value)), state.inheritedValue.raw_value);
             handleSubmit('', state.attribute.id);
             return;
         }
 
-        onChange(rangePickerDates, ...antOnChangeParams);
+        onChange(datePickerDate, ...antOnChangeParams);
 
         // TODO : validate form with await form.validateFields(state.attribute.id)
-        if (state.formElement.settings.required && rangePickerDates === null) {
+        if (state.formElement.settings.required && datePickerDate === null) {
             return;
         }
 
-        const datesToSave = {from: null, to: null};
-        if (rangePickerDates !== null) {
-            const [dateFrom, dateTo] = rangePickerDates;
-            datesToSave.from = String(dateFrom.unix());
-            datesToSave.to = String(dateTo.unix());
+        let dateToSave = null;
+        if (datePickerDate !== null) {
+            dateToSave = String(datePickerDate.unix());
         }
 
-        handleSubmit(datesToSave, state.attribute.id);
+        handleSubmit(dateToSave, state.attribute.id);
     };
 
     const label = localizedTranslation(state.formElement.settings.label, availableLangs);
 
     return (
-        <KitDatePickerRangePickerStyled
+        <KitDatePickerStyled
             value={value}
             onChange={_handleDateChange}
             label={label}
@@ -95,10 +84,7 @@ export const DSRangePicker: FunctionComponent<IDSRangePickerProps> = ({
             helper={
                 state.isInheritedOverrideValue
                     ? t('record_edition.inherited_input_helper', {
-                          inheritedValue: t('record_edition.date_range_from_to', {
-                              from: state.inheritedValue.value.from,
-                              to: state.inheritedValue.value.to
-                          })
+                          inheritedValue: state.inheritedValue.value
                       })
                     : undefined
             }
