+++ libs/ui/src/components/RecordEdition/EditRecordContent/uiElements/StandardField/StandardFieldValue/DSInputWrapper.tsx
@@ -1,22 +1,22 @@
 // Copyright LEAV Solutions 2017
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {KitInputNumber} from 'aristid-ds';
-import {ComponentPropsWithRef, FocusEvent, FunctionComponent, useState} from 'react';
+import {KitInput} from 'aristid-ds';
+import {ChangeEvent, FocusEvent, FunctionComponent, ReactNode, useState} from 'react';
 import {
     IStandardFieldReducerState,
     IStandardFieldValue
 } from '../../../reducers/standardFieldReducer/standardFieldReducer';
-import {Form, InputNumberProps} from 'antd';
+import {Form, InputProps} from 'antd';
 import {IProvidedByAntFormItem} from '_ui/components/RecordEdition/EditRecordContent/_types';
-import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import styled from 'styled-components';
+import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
 import {useValueDetailsButton} from '_ui/components/RecordEdition/EditRecordContent/shared/ValueDetailsBtn/useValueDetailsButton';
 import {RecordFormAttributeFragment} from '_ui/_gqlTypes';
-import {localizedTranslation} from '@leav/utils';
 import {useLang} from '_ui/hooks';
+import {localizedTranslation} from '@leav/utils';
 
-interface IDSInputProps extends IProvidedByAntFormItem<InputNumberProps> {
+interface IDSInputWrapperProps extends IProvidedByAntFormItem<InputProps> {
     state: IStandardFieldReducerState;
     attribute: RecordFormAttributeFragment;
     fieldValue: IStandardFieldValue;
@@ -24,14 +24,11 @@ interface IDSInputProps extends IProvidedByAntFormItem<InputNumberProps> {
     shouldShowValueDetailsButton?: boolean;
 }
 
-const KitInputNumberStyled = styled(KitInputNumber)<{$shouldHighlightColor: boolean}>`
-    .ant-input-number-input-wrap .ant-input-number-input {
-        color: ${({$shouldHighlightColor}) =>
-            $shouldHighlightColor ? 'var(--general-colors-primary-400)' : 'initial'};
-    }
+const KitInputStyled = styled(KitInput)<{$shouldHighlightColor: boolean}>`
+    color: ${({$shouldHighlightColor}) => ($shouldHighlightColor ? 'var(--general-colors-primary-400)' : 'initial')};
 `;
 
-export const DSInputNumber: FunctionComponent<IDSInputProps> = ({
+export const DSInputWrapper: FunctionComponent<IDSInputWrapperProps> = ({
     value,
     onChange,
     state,
@@ -40,19 +37,14 @@ export const DSInputNumber: FunctionComponent<IDSInputProps> = ({
     handleSubmit,
     shouldShowValueDetailsButton = false
 }) => {
-    if (!onChange) {
-        throw Error('DSInputNumber should be used inside a antd Form.Item');
-    }
-
     const {t} = useSharedTranslation();
-    const {lang} = useLang();
     const {errors} = Form.Item.useStatus();
     const {onValueDetailsButtonClick} = useValueDetailsButton({
         value: fieldValue?.value,
         attribute
     });
-
     const [hasChanged, setHasChanged] = useState(false);
+    const {lang: availableLang} = useLang();
 
     const _resetToInheritedValue = () => {
         setHasChanged(false);
@@ -69,18 +61,23 @@ export const DSInputNumber: FunctionComponent<IDSInputProps> = ({
         if (hasChanged || !state.isInheritedValue) {
             handleSubmit(valueToSubmit, state.attribute.id);
         }
-        onChange(valueToSubmit);
+        onChange(event);
     };
 
-    const _handleOnChange: ComponentPropsWithRef<typeof KitInputNumberStyled>['onChange'] = inputValue => {
+    const _handleOnChange = (event: ChangeEvent<HTMLInputElement>) => {
         setHasChanged(true);
-        onChange(inputValue);
+        const inputValue = event.target.value;
+        if (state.isInheritedValue && inputValue === '' && event.type === 'click') {
+            _resetToInheritedValue();
+            return;
+        }
+        onChange(event);
     };
 
-    const label = localizedTranslation(state.formElement.settings.label, lang);
+    const label = localizedTranslation(state.formElement.settings.label, availableLang);
 
     return (
-        <KitInputNumberStyled
+        <KitInputStyled
             required={state.formElement.settings.required}
             label={label}
             onInfoClick={shouldShowValueDetailsButton ? onValueDetailsButtonClick : null}
@@ -93,9 +90,10 @@ export const DSInputNumber: FunctionComponent<IDSInputProps> = ({
                     : undefined
             }
             value={value}
-            onChange={_handleOnChange}
             disabled={state.isReadOnly}
+            allowClear={!state.isInheritedNotOverrideValue}
             onBlur={_handleOnBlur}
+            onChange={_handleOnChange}
             $shouldHighlightColor={!hasChanged && state.isInheritedNotOverrideValue}
         />
     );
